@@ -1,23 +1,23 @@
 name: flutter_acrcloud
 description: A Flutter plugin for the ACRCloud music recognition API.
-version: 0.0.2
+version: 0.1.0
 homepage: https://github.com/nrubin29/flutter_acrcloud
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  json_annotation: ^3.1.1
-  rxdart: ^0.25.0
+  json_annotation: ^4.0.0-nullsafety.0
+  rxdart: ^0.26.0-nullsafety.1
 
 dev_dependencies:
-  build_runner: ^1.10.11
+  build_runner: ^1.11.1
   flutter_test:
     sdk: flutter
-  json_serializable: ^3.5.1
+  json_serializable: ^4.0.0-nullsafety.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
